sm8550-common: Drop qvr/qxr
Does not seem that useful, and conflicts with the
libgrpc++_unsecure package if elf checks are turned
on in extract files.
The libgrpc++_unsecure package builds into the
libgrpc++_unsecure-1.61.1.so binary, but the QVR blob
needs libgrpc++_unsecure.so.
Rather than mess with extract files to fix it
some more, just get rid of it.

Change-Id: Ifebdecf758eb0fff31db6da4a9a6d976e638869e